Extract constants depending on sdk into separate file
To avoid unnecessary dependencies
