+++ src/lib/constants.js
@@ -1,33 +1,8 @@
-import BigNumber from 'bignumber.js';
-import { TxBuilder } from '@aeternity/aepp-sdk/es';
-
 export const MAGNITUDE = 18;
 export const MAGNITUDE_EXA = 18;
 export const MAGNITUDE_GIGA = 9;
 export const MAGNITUDE_MICRO = -6;
 export const MAGNITUDE_PICO = -12;
-
-const STUB_ADDRESS = 'ak_enAPooFqpTQKkhJmU47J16QZu9HbPQQPwWBVeGnzDbDnv9dxp';
-const MAX_UINT256 = BigNumber(2).exponentiatedBy(256).minus(1);
-const MIN_SPEND_TX_FEE_STRING = TxBuilder.calculateMinFee(
-  'spendTx', {
-    params: {
-      senderId: STUB_ADDRESS,
-      recipientId: STUB_ADDRESS,
-      amount: MAX_UINT256,
-      ttl: MAX_UINT256,
-      nonce: MAX_UINT256,
-    },
-  },
-);
-
-export const MIN_SPEND_TX_FEE = BigNumber(MIN_SPEND_TX_FEE_STRING).shiftedBy(-MAGNITUDE);
-export const MAX_REASONABLE_FEE = MIN_SPEND_TX_FEE.multipliedBy(10);
-
-const toMicro = value => value.shiftedBy(-MAGNITUDE_MICRO).toFixed();
-
-export const MIN_SPEND_TX_FEE_MICRO = toMicro(MIN_SPEND_TX_FEE);
-export const MAX_REASONABLE_FEE_MICRO = toMicro(MAX_REASONABLE_FEE);
 
 export const AENS_DOMAIN = '.chain';
 export const MAX_AUCTION_NAME_LENGTH = 12 + AENS_DOMAIN.length;
